chore: rename index.ts -> mod.ts
