+++ src/make_test.ts
@@ -1,4 +1,4 @@
-import { def, get, make } from "./index.ts";
+import { def, get, make } from "./mod.ts";
 import { Foo } from "./test_fixture.ts";
 import { assert, assertEquals, clearComponents } from "./test_helper.ts";
